+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Generator configuration by build property
 - Added EmbeddingFileNameType config
+- SourceExpander.Console: Resolve dependency
 
 ## [4.1.1] - 2022-02-26
 ### Changed